app/spark/requirements.txt
pyspark==3.2.0
kafka-python==2.0.2
pandas==1.3.0
numpy==1.21.0
pymongo==4.6.1
xgboost==1.7.6
scikit-learn==1.0.2
joblib==1.2.0 